[#1]refactor(domain): entity 패키지 삭제
