+++ src/main/java/baksakcci/shoppingmall/order/application/PlaceOrderServiceImpl.java
@@ -5,8 +5,8 @@ import baksakcci.shoppingmall.catalog.domain.repository.ProductRepository;
 import baksakcci.shoppingmall.order.application.dto.OrderCreate;
 import baksakcci.shoppingmall.order.application.dto.OrderItemCreate;
 import baksakcci.shoppingmall.order.application.port.PlaceOrderService;
-import baksakcci.shoppingmall.order.domain.entity.Order;
-import baksakcci.shoppingmall.order.domain.entity.OrderItem;
+import baksakcci.shoppingmall.order.domain.Order;
+import baksakcci.shoppingmall.order.domain.OrderItem;
 import baksakcci.shoppingmall.order.application.port.OrderRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
