+++ src/main/java/baksakcci/shoppingmall/order/application/port/PlaceOrderService.java
@@ -1,7 +1,7 @@
 package baksakcci.shoppingmall.order.application.port;
 
 import baksakcci.shoppingmall.order.application.dto.OrderCreate;
-import baksakcci.shoppingmall.order.domain.entity.Order;
+import baksakcci.shoppingmall.order.domain.Order;
 
 public interface PlaceOrderService {
