+++ src/main/java/baksakcci/shoppingmall/order/domain/Order.java
@@ -1,4 +1,4 @@
-package baksakcci.shoppingmall.order.domain.entity;
+package baksakcci.shoppingmall.order.domain;
 
 import baksakcci.shoppingmall.order.application.dto.OrderCreate;
 import java.time.LocalDateTime;
