+++ src/main/java/baksakcci/shoppingmall/order/domain/OrderState.java
@@ -1,4 +1,4 @@
-package baksakcci.shoppingmall.order.domain.entity;
+package baksakcci.shoppingmall.order.domain;
 
 public enum OrderState {
     // 결제 대기, 상품 준비, 배송 대기, 배송중, 배송 완료
