+++ src/main/java/baksakcci/shoppingmall/order/domain/Orderer.java
@@ -1,0 +1,7 @@
+package baksakcci.shoppingmall.order.domain;
+
+public class Orderer {
+
+    private Long id;
+
+}
