--- src/main/java/baksakcci/shoppingmall/order/domain/entity/Orderer.java
@@ -1,7 +1,0 @@
-package baksakcci.shoppingmall.order.domain.entity;
-
-public class Orderer {
-
-    private Long id;
-
-}
