+++ src/main/java/baksakcci/shoppingmall/order/infra/OrderEntity.java
@@ -1,9 +1,9 @@
 package baksakcci.shoppingmall.order.infra;
 
-import baksakcci.shoppingmall.order.domain.entity.DeliveryInfo;
-import baksakcci.shoppingmall.order.domain.entity.Order;
-import baksakcci.shoppingmall.order.domain.entity.OrderItem;
-import baksakcci.shoppingmall.order.domain.entity.OrderState;
+import baksakcci.shoppingmall.order.domain.DeliveryInfo;
+import baksakcci.shoppingmall.order.domain.Order;
+import baksakcci.shoppingmall.order.domain.OrderItem;
+import baksakcci.shoppingmall.order.domain.OrderState;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
