+++ src/main/java/baksakcci/shoppingmall/order/infra/OrderFakeRepository.java
@@ -1,12 +1,11 @@
 package baksakcci.shoppingmall.order.infra;
 
-import baksakcci.shoppingmall.order.domain.entity.Order;
+import baksakcci.shoppingmall.order.domain.Order;
 import baksakcci.shoppingmall.order.application.port.OrderRepository;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Optional;
 
 public class OrderFakeRepository implements OrderRepository {
