+++ src/main/java/baksakcci/shoppingmall/order/infra/OrderRepositoryImpl.java
@@ -4,8 +4,8 @@ import baksakcci.shoppingmall.catalog.domain.entity.Product;
 import baksakcci.shoppingmall.catalog.infra.ProductEntity;
 import baksakcci.shoppingmall.catalog.infra.ProductJpaRepository;
 import baksakcci.shoppingmall.order.application.port.OrderRepository;
-import baksakcci.shoppingmall.order.domain.entity.Order;
-import baksakcci.shoppingmall.order.domain.entity.OrderItem;
+import baksakcci.shoppingmall.order.domain.Order;
+import baksakcci.shoppingmall.order.domain.OrderItem;
 import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
