+++ src/test/java/baksakcci/shoppingmall/order/application/PlaceOrderServiceTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import baksakcci.shoppingmall.catalog.domain.entity.Product;
 import baksakcci.shoppingmall.catalog.infra.ProductFakeRepository;
-import baksakcci.shoppingmall.order.domain.entity.Order;
+import baksakcci.shoppingmall.order.domain.Order;
 import baksakcci.shoppingmall.order.application.dto.OrderCreate;
-import baksakcci.shoppingmall.order.domain.entity.OrderItem;
+import baksakcci.shoppingmall.order.domain.OrderItem;
 import baksakcci.shoppingmall.order.application.dto.OrderItemCreate;
 import baksakcci.shoppingmall.order.infra.OrderFakeRepository;
 import java.time.LocalDate;
